+++ src/test/java/io/hhplus/concert/application/waitingqueue/WaitingQueueFacadeIntegrationTest.java
@@ -3,8 +3,8 @@ package io.hhplus.concert.application.waitingqueue;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
-import io.hhplus.concert.application.waitingqueue.dto.WaitingQueueDto.WaitingQueueInfo;
-import io.hhplus.concert.application.waitingqueue.dto.WaitingQueueDto.WaitingQueueWithOrderInfo;
+import io.hhplus.concert.application.waitingqueue.WaitingQueueDto.WaitingQueueInfo;
+import io.hhplus.concert.application.waitingqueue.WaitingQueueDto.WaitingQueueWithOrderInfo;
 import io.hhplus.concert.domain.common.ServicePolicy;
 import io.hhplus.concert.domain.waitingqueue.exception.WaitingQueueErrorCode;
 import io.hhplus.concert.domain.waitingqueue.exception.WaitingQueueException;
